docker/kafka_apps/4_kafka_connect/seeders/requirements.txt
psycopg2-binary
sqlalchemy
cassandra-driver
faker